Prune variable
The current script holds the old token and payerid in its variable, which result in error if more than one payment is made. (console.log(plans[ req.body.plan ]) in the start of the /checkout route and make two payments, which another is succeeded)

This patch removes those two variables, resulting in working code.